move docker push
